third_party/WebKit/LayoutTests/fullscreen/api/element-request-fullscreen-and-remove-iframe.html

<!DOCTYPE html>
<title>Element#requestFullscreen() in iframe followed by removing the iframe</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<iframe allowfullscreen></iframe>
<script>
// Adapted from https://github.com/w3c/web-platform-tests/pull/4250
// TODO(foolip): Remove this test when the above is imported and passing.
async_test(t => {
  const iframe = document.querySelector("iframe");
  const iframeDocument = iframe.contentDocument;

  document.onfullscreenchange = t.step_func_done(() => {
    assert_equals(document.fullscreenElement, null);
    assert_equals(iframeDocument.fullscreenElement, null);
  });
  document.onfullscreenerror = t.unreached_func("fullscreenerror event");
  iframeDocument.onfullscreenchange = t.unreached_func("iframe fullscreenchange event");
  iframeDocument.onfullscreenerror = t.unreached_func("iframe fullscreenerror event");

  trusted_click(t.step_func(() => {
    iframeDocument.body.requestFullscreen();
    iframe.remove();
  }), document.body);
});
</script>
